Imported Upstream version 0.6.0

+++ lib/unix_utils.h
@@ -13,7 +13,7 @@ static inline char * get_bstr(value v_bstr, value v_pos)
   return (char *) Caml_ba_data_val(v_bstr) + Long_val(v_pos);
 }
 
-static inline struct iovec * copy_iovecs(long *total_len, value v_iovecs, int n)
+static inline struct iovec * copy_iovecs(size_t *total_len, value v_iovecs, int n)
 {
   struct iovec *iovecs = caml_stat_alloc(sizeof(struct iovec) * n);
   for (--n; n >= 0; --n) {
@@ -21,7 +21,7 @@ static inline struct iovec * copy_iovecs(long *total_len, value v_iovecs, int n)
     value v_iovec = Field(v_iovecs, n);
     value v_iov_base = Field(v_iovec, 0);
     value v_iov_pos = Field(v_iovec, 1);
-    int iov_len = Long_val(Field(v_iovec, 2));
+    size_t iov_len = Long_val(Field(v_iovec, 2));
     iovec->iov_len = iov_len;
     *total_len += iov_len;
     iovec->iov_base = get_bstr(v_iov_base, v_iov_pos);
